VRF - Contract Address section regarding Polygon
This pull request purpose is to inform developers that the LINK token address to be used for the VRF on the Polygon Network is not the same as the LINK token obtained through dexes like Quickswap or through bridging from L1 to L2. Since the common LINk token on Polygon is not ERC-677 compatible, it cannot be used to call the VRF. The present change is duplicating the information taken from https://docs.chain.link/docs/link-token-contracts/#polygon-matic and placing it on the VRF - Contract Address section.