Update README.md
Dear Maintainers,

I have made some changes to the README file that I believe improve the overall clarity and effectiveness of the documentation introduction. Specifically, I added a link to the Chainlink developer documentation website and included an invitation for users to start building with Chainlink.

Please let me know if there are any further changes or adjustments that need to be made before merging this pull request.

Thank you for your time and consideration.